+++ resolwe/flow/utils/__init__.py
@@ -17,8 +17,8 @@ import hashlib
 
 def get_data_checksum(proc_input, proc_slug, proc_version):
     """Compute checksum of processor inputs, name and version."""
-    hash_ = hashlib.sha1()
-    hash_.update(str(proc_input).encode('utf-8'))
-    hash_.update(proc_slug.encode('utf-8'))
-    hash_.update(str(proc_version).encode('utf-8'))
-    return hash_.hexdigest()
+    checksum = hashlib.sha1()
+    checksum.update(str(proc_input).encode('utf-8'))
+    checksum.update(proc_slug.encode('utf-8'))
+    checksum.update(str(proc_version).encode('utf-8'))
+    return checksum.hexdigest()